Future Prep
created empty can attack function that determines if a pokemon can
attack (para, con, etc. will go here)
created empty endTurnEffects that gets called at the end of every turn
(leech seed, poison, etc. will go here)
associated an array of strings with moves that contain properties that
need to be taken into account when the move is called (weather,
priority, etc.)
reset stat changes on switches
created a bp variable so that it is easier to change if need be